Move tests to match new namespace
